System call implementation
This documentation contains the description and implementation steps of the system call "chown" using the c programming language. It lists the required header file, the correct syntax, the expected return type and its values. It also shows how the implementation on Minix OS looks like. In Minix, chown can be used directly in the command line interface by any user with the right privileges or  by using it inside a c code and here we only used it inside a c code, and both of them have different syntax structures. The attached c code shows a correct usage of the system call on the right user and group and exception handling in case of errors.